xpon-portable: sync latest GPON/serdes/OMCI development
Sync the package/kernel/xpon-portable driver with the latest local
development (superset of the prior EPON/XEPON work already on this
branch). Key additions over the previous xpon-portable state:

- GPON bring-up path (xpon_gpon.c) and G.984 PLOAM/OMCI plumbing
- serdes bring-up: JCPLL/TXPLL/Phya PLL sequences + xsgmii_ini init
  (pon_serdes_init, complementing mainline phy-airoha-xpon configure)
- OMCI/OMCC helpers (apply/control/software/event) + GEM sniffer at O5
- XGS-PON register map alignment to real firmware regmap
- Act/Netdev/QDMA refinements; include/xpon.h + xpon_ioctl.h extended

Local 18-commit history is preserved in the standalone xpon-portable
repo; this commit brings the code current in the OpenWrt tree.

+++ package/kernel/xpon-portable/src/core/xpon_omcc.c
@@ -114,9 +114,9 @@ MODULE_PARM_DESC(omcc_loopback,
 #define RD_BE32(p) (((u32)(p)[0] << 24) | ((u32)(p)[1] << 16) | \
 		    ((u32)(p)[2] << 8)  |  (u32)(p)[3])
 #define RD_BE16(p) (((u32)(p)[0] << 8)  |  (u32)(p)[1])
-#define WR_BE32(p, v) do { (p)[0] = (v) >> 24; (p)[1] = (v) >> 16; \
-			   (p)[2] = (v) >> 8;  (p)[3] = (v); } while (0)
-#define WR_BE16(p, v) do { (p)[0] = (v) >> 8; (p)[1] = (v); } while (0)
+#define WR_BE32(p, v) do { (p)[0] = (u8)((v) >> 24); (p)[1] = (u8)((v) >> 16); \
+			   (p)[2] = (u8)((v) >> 8);  (p)[3] = (u8)(v); } while (0)
+#define WR_BE16(p, v) do { (p)[0] = (u8)((v) >> 8); (p)[1] = (u8)(v); } while (0)
 
 /*
  * CRC-32/ITU-I.363.5 (non-reflected), the G.988 OMCI MIC.
